Restructure ES|QL ComputeService (#120312)
This change divides the ComputeService into three smaller components:

- ComputeService, responsible for managing high-level computes

- ClusterComputeHandler, responsible for handling cluster-level computes 
  across multiple clusters

- DataNodeComputeHandler, responsible for managing data-node-level 
  computes within a single cluster

This restructuring simplifies the management of these responsibilities.
